refactor(resultMessge): 결과 출력하는 함수 분리
