+++ src/App.js
@@ -1,17 +1,11 @@
-import { Console } from '@woowacourse/mission-utils'
+import { Console } from '@woowacourse/mission-utils';
+import { getResultMessages } from './resultMessages.js';
 import LottoGame from "./LottoGame.js";
 import InputHandler from "./InputHandler.js";
 
 const printWinningResults = (lottoGame, userLotto, winningResults) => {
-  const resultMessages = [
-    `3개 일치 (5,000원) - ${winningResults[3]}개`,
-    `4개 일치 (50,000원) - ${winningResults[4]}개`,
-    `5개 일치 (1,500,000원) - ${winningResults[5]}개`,
-    `5개 일치, 보너스 볼 일치 (30,000,000원) - ${winningResults['5+bonus']}개`,
-    `6개 일치 (2,000,000,000원) - ${winningResults[6]}개`
-  ]
-
-  Console.print("당첨 통계");
+  const resultMessages = getResultMessages(winningResults);
+  Console.print("\n당첨 통계");
   Console.print("---");
 
   const profitRate = lottoGame.calculateProfit(userLotto);
@@ -28,10 +22,11 @@ class App {
     const purchasePrice = await InputHandler.getPurchasePrice();
     lottoGame.setPurchasePrice(purchasePrice);
 
-    Console.print(`${lottoGame.calculateLottoCount()}개를 구매했습니다.`);
+    Console.print(`\n${lottoGame.calculateLottoCount()}개를 구매했습니다.`);
 
     const userLotto = lottoGame.generateLotto();
     userLotto.forEach((lotto) => Console.print(lotto.toString()));
+    Console.print('');
 
     const winNumbers = await InputHandler.getWinNumbers();
     lottoGame.setWinNumbers(winNumbers);
